Add InMemoryDatabase

+++ scripts/utils.py
@@ -8,23 +8,25 @@ import numpy as np
 from pathlib import Path
 from tqdm import tqdm
 from sklearn.linear_model import LinearRegression
-from typing import Callable, Iterable, Sequence
+from typing import Callable
 from sklearn.model_selection import KFold
-from retry import retry
+
+import dask
+from multiprocessing import Pool
 
 from functools import wraps
 from time import time
 
 
 def timing(f):
-
-    @retry(Exception, tries=1, delay=1)
     @wraps(f)
     def wrap(*args, **kw):
+        print(f"started {f.__name__}")
+
         ts = time()
         result = f(*args, **kw)
         te = time()
-        print("func:%r took: %2.4f sec" % (f.__name__, te - ts))
+        print(f"  running {f.__name__} took {te-ts:2.4f} sec")
         return result
 
     return wrap
@@ -35,6 +37,106 @@ class RawDataPath:
     AmpC = f"{raw_data}/AmpC/"
     D4 = f"{raw_data}/D4"
     D4_small = f"{raw_data}/D4_small"
+    D4_medium = f"{raw_data}/D4_medium"
+
+
+class InMemoryDatabase:
+    def __init__(
+        self,
+        which: str,
+        chunksize: int = 1_000_000,
+        n_bits: int = 10,
+        dropna: bool = True,
+        n_workers: int = 32,
+    ):
+        if which.lower() == "d4":
+            path = RawDataPath.D4
+        elif which.lower() == "ampc":
+            path = RawDataPath.AmpC
+        elif which.lower() == "d4_small":
+            path = RawDataPath.D4_small
+        elif which.lower() == "d4_medium":
+            path = RawDataPath.D4_medium
+        else:
+            raise ValueError(f"wrong dataset name: {which=}")
+
+        self.path = Path(path).resolve()
+        self.which = which
+        self.raw = f"{self.path}/raw.csv"
+        self.prepared = f"{self.path}/prepared.parquet"
+        self.chunksize = chunksize
+        self.dropna = dropna
+        self.n_bits = n_bits
+        self.n_workers = n_workers
+        self.compute_kwargs = {
+            "n_workers": self.n_workers,
+            "scheduler": "processes",
+            "threads_per_worker":4,
+            "chunksize": 1,
+        }
+
+        if not Path(self.prepared).exists():
+            self.read_raw_data()
+            self.add_fingerprints()
+            self.serialize()
+        else:
+            self.df_raw = self.read_prepared()
+
+    @timing
+    def read_raw_data(self):
+        df = pd.read_csv(self.raw, engine="pyarrow")
+        if self.dropna:
+            df["dockscore"] = pd.to_numeric(df["dockscore"], errors="coerce")
+            df = df.dropna().reset_index()
+        self.df_raw = df
+        return df
+
+    @timing
+    def read_prepared(self):
+        df = pd.read_parquet(self.prepared)
+        return df
+
+    @timing
+    def read_column(self, column: str, idx: np.ndarray = None) -> np.ndarray:
+        df = self.df_raw
+        if idx is not None:
+            df = df[df.index.isin(idx)]
+        if column == "fingerprints":
+            columns = [col for col in self.df_raw.columns if col.startswith("fps")]
+            return df[columns].values
+        else:
+            return df[column].values
+
+    @timing
+    def add_fingerprints(self, column="smiles", force: bool = False):
+        if force or not Path(self.prepared).exists():
+            all_smiles = self.read_column(column=column)
+            n_splits = 10 + len(all_smiles) // self.chunksize
+            tasks = np.array_split(all_smiles, n_splits)
+            tasks = delayed(
+                [
+                    _fingerprint_helper_inplace(task, fpsize=2**self.n_bits)
+                    for task in tasks
+                ]
+            )
+
+            results = np.vstack(dask.compute(tasks, **self.compute_kwargs)[0])
+            colnames = [f"fps_{i}" for i in range(results.shape[1])]
+            self.df_raw[colnames] = results
+
+    @timing
+    def get_random_batch(self, batchsize: int) -> tuple[np.ndarray, np.ndarray]:
+        scores = self.read_column("dockscore")
+        idx = np.arange(len(scores))
+        np.random.shuffle(idx)
+        rv_idx = idx[:batchsize]
+        rv_scores = scores[rv_idx]
+        return rv_idx, rv_scores
+
+    @timing
+    def serialize(self, force: bool = False):
+        if not Path(self.prepared).exists() or force:
+            self.df_raw.to_parquet(self.prepared)
 
 
 class Database:
@@ -43,7 +145,9 @@ class Database:
         which: str,
         chunksize: int = 50_000,
         initialize: bool = False,
-        dropna: bool = True
+        n_bits: int = 10,
+        dropna: bool = True,
+        **compute_kwargs,
     ):
         if which.lower() == "d4":
             path = RawDataPath.D4
@@ -59,6 +163,15 @@ class Database:
         self.raw = f"{self.path}/raw.csv"
         self.chunksize = chunksize
         self.dropna = dropna
+        self.n_bits = n_bits
+
+        if not compute_kwargs:
+            compute_kwargs = {
+                "n_workers": 32,
+                "threads_per_worker": 4,
+                "scheduler": "processes",
+            }
+        self.compute_kwargs = compute_kwargs
 
         if initialize:
             self.read_raw_data()
@@ -66,11 +179,12 @@ class Database:
             self.add_fingerprints()
 
     @timing
+    @functools.lru_cache(maxsize=1)
     def read_raw_data(self):
         df = pd.read_csv(self.raw, engine="pyarrow")
         if self.dropna:
-            df['dockscore'] = pd.to_numeric(df['dockscore'], errors='coerce')
-            df = df.dropna()
+            df["dockscore"] = pd.to_numeric(df["dockscore"], errors="coerce")
+            df = df.dropna().reset_index()
         self.df_raw = df
         return df
 
@@ -99,12 +213,19 @@ class Database:
                     for idx, chunk in enumerate(np.array_split(df, n_splits))
                 ]
             )
-            # computations.compute(n_workers=48, scheduler='processes')
-            computations.compute()
+            computations.compute(**self.compute_kwargs)
 
     @timing
     def get_filenames_for(self, column: str) -> list[Path]:
-        return tuple(sorted([p for p in Path(f"{self.path}/{column}").iterdir() if p.suffix == '.parquet']))
+        return tuple(
+            sorted(
+                [
+                    p
+                    for p in Path(f"{self.path}/{column}").iterdir()
+                    if p.suffix == ".parquet"
+                ]
+            )
+        )
 
     @timing
     def read_column(self, column: str, idx: np.ndarray = None) -> np.ndarray:
@@ -119,8 +240,7 @@ class Database:
         func = delayed(func)
         batches = self.get_filenames_for(column)
         tasks = delayed([func(batch) for batch in batches])
-        # results = tasks.compute(n_workers=48, scheduler='processes')
-        results = tasks.compute()
+        results = tasks.compute(**self.compute_kwargs)
         return agg(results)
 
     @timing
@@ -137,8 +257,7 @@ class Database:
                 for batch in self.get_filenames_for(column)
             ]
         )
-        # tasks.compute(n_workers=64, scheduler='processes')
-        tasks.compute()
+        tasks.compute(**self.compute_kwargs)
 
     def _execute(
         self, df_func: Callable, agg_func: Callable, n_parts: int = 12, columns=None
@@ -149,8 +268,7 @@ class Database:
                 df_func(self._read_df(filename, columns=columns))
                 for filename in self.parquet_files
             ]
-        ).compute()
-        # ).compute(n_workers=n_parts, scheduler='processes')
+        ).compute(**self.compute_kwargs)
         return agg_func(results)
 
     @timing
@@ -171,13 +289,14 @@ def _replace_column_name(p: Path, target: str) -> Path:
 def _fingerprint_helper(
     input_filename: str,
     output_filename: str,
+    fpsize: int = 2048,
     column: str = "smiles",
     force: bool = False,
 ):
     if force or not Path(output_filename).exists():
-        generator = rdFingerprintGenerator.GetMorganGenerator(radius=2, fpSize=2048)
+        generator = rdFingerprintGenerator.GetMorganGenerator(radius=2, fpSize=fpsize)
         smiles2fp = lambda s: generator.GetFingerprintAsNumPy(Chem.MolFromSmiles(s))
-        df = pd.read_parquet(input_filename)
+        df = pd.read_parquet(input_filename, engine="fastparquet")
         arr = np.vstack([smiles2fp(s) for s in df[column]])
         arr_as_df = pd.DataFrame(
             arr, columns=[f"fp_{idx:04}" for idx in range(arr.shape[1])], index=df.index
@@ -185,41 +304,49 @@ def _fingerprint_helper(
         arr_as_df.to_parquet(output_filename)
 
 
+@delayed
+def _fingerprint_helper_inplace(
+    arr: list[str],
+    fpsize,
+):
+    generator = rdFingerprintGenerator.GetMorganGenerator(radius=2, fpSize=fpsize)
+    smiles2fp = lambda s: generator.GetFingerprintAsNumPy(Chem.MolFromSmiles(s))
+    return np.vstack([smiles2fp(s) for s in arr])
+
+
 def _read_column_with_idx(filename: str, column: str, idx: np.ndarray) -> np.ndarray:
-    df = pd.read_parquet(filename)
+    df = pd.read_parquet(filename, engine="fastparquet")
     if idx is not None:
         df = df[df.index.isin(idx)]
     return df[column].values
 
 
 def _read_fingerprints(filename: str, idx: np.ndarray) -> np.ndarray:
-    df = pd.read_parquet(filename)
+    df = pd.read_parquet(filename, engine="fastparquet")
     if idx is not None:
         df = df[df.index.isin(idx)]
     return df.values
 
 
 def apply_model_to(fingerprints_filename: str, model: "OneShotModel") -> np.ndarray:
-    X = pd.read_parquet(fingerprints_filename, engine='fastparquet').values
+    X = pd.read_parquet(fingerprints_filename, engine="fastparquet").values
     y = model.predict(X)
-    with open(f'{fingerprints_filename}.done', 'w') as fout:
-        print('DONE', file=fout)
+    with open(f"{fingerprints_filename}.done", "w") as fout:
+        print("DONE", file=fout)
     return y
+
+
+def apply_model_to_ndarray(X: np.ndarray, model: "OneShotModel") -> np.ndarray:
+    return model.predict(X)
 
 
 @functools.lru_cache(maxsize=20)
 def apply_single_model_to(
-    filenames: tuple[str], model, n_parts: int = 48
-) ->np.ndarray:
+    filenames: tuple[str],
+    model,
+) -> np.ndarray:
     computations = delayed([apply_model_to(filename, model) for filename in filenames])
-    # ypreds = computations.compute(scheduler="processes", n_workers=n_parts)
-    ypreds = computations.compute()
-
-    # from multiprocessing import Pool
-    # from functools import partial
-    # with Pool(n_processes=n_parts) as pool:
-    #     ypreds = pool.apply(partial(apply_model_to, model=model), filenames)
-
+    ypreds = computations.compute(**model.compute_kwargs)
     return np.hstack(ypreds)
 
 
@@ -242,12 +369,15 @@ def ordinal(x: np.ndarray) -> np.ndarray:
 
 
 class OneShotModel:
-    def __init__(self, n_splits: int = 5, regressor_factory: Callable = None):
+    def __init__(
+        self, n_splits: int = 5, regressor_factory: Callable = None, **compute_kwargs
+    ):
         self.n_splits = n_splits
 
         if regressor_factory is None:
             regressor_factory = lambda: LinearRegression(n_jobs=4)
         self._regressor_factory = regressor_factory
+        self.compute_kwargs = compute_kwargs
 
     def fit(self, X: np.ndarray, y: np.ndarray):
         kf = KFold(n_splits=self.n_splits)
@@ -257,8 +387,9 @@ class OneShotModel:
                 for idx, _ in kf.split(X, y)
             ]
         )
-        # models = models.compute(n_workers=self.n_splits)
-        models = models.compute()
+        models = models.compute(
+            n_workers=self.n_splits, scheduler="processes", chunksize=1
+        )
         self.models = models
         return self
 
@@ -275,6 +406,8 @@ class ActiveLearningModel:
         self,
         regime: str = "LastModel",
         oneshot_factory: Callable = None,
+        memory_size: int = 20,
+        **compute_kwargs,
     ):
         if regime not in (
             "LastModel",
@@ -283,13 +416,22 @@ class ActiveLearningModel:
             raise ValueError(f"Wrong {regime=}")
         self.regime = regime
 
+        if not compute_kwargs:
+            compute_kwargs = {
+                "n_workers": 32,
+                "threads_per_worker": 1,
+                "scheduler": "processes",
+            }
+        self.compute_kwargs = compute_kwargs
+
         if oneshot_factory is None:
-            oneshot_factory = lambda: OneShotModel()
+            oneshot_factory = lambda: OneShotModel(compute_kwargs=compute_kwargs)
         self._oneshot_factory = oneshot_factory
 
         self.models: dict[int, object] = {}
         self.seen_molecules: dict[int, np.ndarray] = {}
         self.iteration = 0
+        self.memory_size = memory_size
 
     @timing
     def add_iteration(
@@ -318,10 +460,12 @@ class ActiveLearningModel:
 
         if ignore_seen:
             preds[seen_molecules] = float("inf")
-        
-        for done in db.get_filenames_for('fingerprints'):
-            if done.suffix == '.done':
-                Path(done).unlink()
+
+        if isinstance(db, Database):
+            for done in db.get_filenames_for("fingerprints"):
+                if done.suffix == ".done":
+                    Path(done).unlink()
+
         return preds
 
     @timing
@@ -334,14 +478,31 @@ class ActiveLearningModel:
         return idx
 
     def _apply_with_lastmodel(self, db: Database) -> np.ndarray:
-        model = self.models[self.iteration - 1]
-        return apply_single_model_to(db.get_filenames_for("fingerprints"), model)
+        if isinstance(db, Database):
+            model = self.models[self.iteration - 1]
+            return apply_single_model_to(db.get_filenames_for("fingerprints"), model)
+        elif isinstance(db, InMemoryDatabase):
+            model = self.models[self.iteration - 1]
+            X = db.read_column("fingerprints")
+            return model.predict(X)
+        else:
+            raise ValueError
 
     def _apply_with_meanrank(self, db: Database) -> np.ndarray:
-        results_for_each_model = [
-            apply_single_model_to(db.get_filenames_for("fingerprints"), model)
-            for model in self.models.values()
-        ]
-        ordinals = [ordinal(arr) for arr in results_for_each_model]
-        return np.vstack(ordinals).mean(axis=0)
-
+        if isinstance(db, Database):
+            results_for_each_model = [
+                apply_single_model_to(db.get_filenames_for("fingerprints"), model)
+                for model in list(self.models.values())[-self.memory_size :]
+            ]
+            ordinals = [ordinal(arr) for arr in results_for_each_model]
+            return np.vstack(ordinals).mean(axis=0)
+        elif isinstance(db, InMemoryDatabase):
+            X = db.read_column("fingerprints")
+            results_for_each_model = [
+                model.predict(X)
+                for model in list(self.models.values())[-self.memory_size :]
+            ]
+            ordinals = [ordinal(arr) for arr in results_for_each_model]
+            return np.vstack(ordinals).mean(axis=0)
+        else:
+            raise ValueError
